+++ front/src/UI/organisms/ChatSendBox/ChatSendBoxQueries.ts
@@ -1,0 +1,13 @@
+import { gql } from '@apollo/client';
+
+export const CREATE_CHAT_LOG = gql`
+  mutation CreateChatLog($user: CreateChatLogInput!) {
+    createChatLog(createChatLogInput: $user) {
+      index
+      chatUUID
+      userID
+      message
+      createdAt
+    }
+  }
+`;
